Safari-609.1.11.1/LayoutTests/platform/ios-12/fast/css/apple-system-colors-expected.txt

-apple-system-blue : rgb(0, 122, 255)
-apple-system-brown : rgba(0, 0, 0, 0)
-apple-system-gray : rgb(142, 142, 147)
-apple-system-green : rgb(76, 217, 100)
-apple-system-orange : rgb(255, 149, 0)
-apple-system-pink : rgb(255, 45, 85)
-apple-system-purple : rgb(88, 86, 214)
-apple-system-red : rgb(255, 59, 48)
-apple-system-yellow : rgb(255, 204, 0)
-apple-wireless-playback-target-active : rgb(0, 122, 255)
current-color with inherited -apple-system-blue : rgb(0, 122, 255)
